ERS 0.59.13: - на экран Magic Service добавлено отображение портов ВГшки, адрес бряка и текущую частоту процессора. - переделана задержка детекта HDD для срабатывания только после включения. - убран детект SLAVE HDD. DOS 0.58: - добавлена расширенная команда TURBO для просмотра/установки частоты процессора. Обновлен асм, бинарики собраны только EXE. Залит его исходный архив.
git-svn-id: svn://localhost/local-pentevo@1120 c6c76956-fad2-4150-a3a9-ea0a427205ba

+++ tools/asl/asl-current/install.sh
@@ -36,9 +36,6 @@ if [ "${BINPATH}" != "" ]; then
    chmod 755 ${BINPATH}/$i${TARG_EXEXTENSION}
   fi
  done
- if test "${TARG_EXEXTENSION}" = ".exe"; then
-  mv ${BINPATH}/asl${TARG_EXEXTENSION} ${BINPATH}/asw${TARG_EXEXTENSION}
- fi
 fi
 
 if test "${TARG_EXEXTENSION}" = ".exe"; then
@@ -84,13 +81,15 @@ fi
 if [ "${LIBPATH}" != "" ]; then
  ${MKDIRHIER} ${LIBPATH}
  chmod 755 ${LIBPATH}
- for file in ${TARG_OBJDIR}*.msg; do
-  base=`basename ${file}`
-  #echo copy ${file} to ${LIBPATH}/${base} ...
-  if cp ${file} ${LIBPATH}/ ; then
-   chmod 644 ${LIBPATH}/${base}
-  fi
- done
+ if [ -f as.msg ]; then
+  for file in ${TARG_OBJDIR}*.msg; do
+   base=`basename ${file}`
+   #echo copy ${file} to ${LIBPATH}/${base} ...
+   if cp ${file} ${LIBPATH}/ ; then
+    chmod 644 ${LIBPATH}/${base}
+   fi
+  done
+ fi
 fi
 
 if [ "${DOCPATH}" != "" ]; then
@@ -111,6 +110,8 @@ if [ "${DOCPATH}" != "" ]; then
  chmod 644 ${DOCPATH}/ps*.tex
  cp doc_COM/biblio.tex ${DOCPATH}
  chmod 644 ${DOCPATH}/biblio.tex
+ cp doc_COM/cp3finst.tex ${DOCPATH}
+ chmod 644 ${DOCPATH}/cp3finst.tex
  cp COPYING ${DOCPATH}
  chmod 644 ${DOCPATH}/COPYING
 fi
